Reorganize files and add more READMEs
