+++ packages/components/src/flex/flex-block/README.md
@@ -1,0 +1,15 @@
+# FlexBlock
+
+A layout component to contain items of a fixed width within `Flex`.
+
+## Usage
+
+See [`flex/README.md#usage`](../flex/README.md#usage) for how to use `FlexBlock`.
+
+## Props
+
+### display
+
+**Type**: `[CSSProperties['display']]`
+
+The CSS display property of `FlexBlock`.
