+++ packages/components/src/flex/flex-block/index.js
@@ -1,0 +1,2 @@
+export { default } from './component';
+export { useFlexBlock } from './hook';
