+++ packages/components/src/flex/flex-item/README.md
@@ -1,0 +1,21 @@
+# FlexItem
+
+A layout component to contain items of a fixed width within `Flex`.
+
+## Usage
+
+See [`flex/README.md#usage`](../flex/README.md#usage) for how to use `FlexItem`.
+
+## Props
+
+### display
+
+**Type**: `[CSSProperties['display']]`
+
+The CSS display property of `FlexItem`.
+
+### isBlock
+
+**Type**: `[boolean]`
+
+Determins if `FlexItem` should render as an adaptive full-width block.
