+++ packages/components/src/flex/flex-item/index.js
@@ -1,0 +1,2 @@
+export { default } from './component';
+export { useFlexItem } from './hook';
