+++ packages/components/src/flex/flex/README.md
@@ -27,7 +27,7 @@ function Example() {
 
 ##### align
 
-**Type**: `CSS['alignItems']`
+**Type**: `CSSProperties['alignItems']`
 
 Aligns children using CSS Flexbox `align-items`. Vertically aligns content if the `direction` is `row`, or horizontally aligns content if the `direction` is `column`.
 
@@ -35,31 +35,31 @@ In the example below, `flex-start` will align the children content to the top.
 
 ##### direction
 
-**Type**: `FlexDirection`
+**Type**: `[ResponsiveCSSValue<CSSProperties['flexDirection']>]`
 
 The direction flow of the children content can be adjusted with `direction`. `column` will align children vertically and `row` will align children horizontally.
 
 ##### expanded
 
-**Type**: `boolean`
+**Type**: `[boolean]`
 
 Expands to the maximum available width (if horizontal) or height (if vertical).
 
 ##### gap
 
-**Type**: `number`
+**Type**: `[number | string]`
 
 Spacing in between each child can be adjusted by using `gap`. The value of `gap` works as a multiplier to the library's grid system (base of `4px`).
 
 ##### justify
 
-**Type**: `CSS['justifyContent']`
+**Type**: `[CSSProperties['justifyContent']]`
 
 Horizontally aligns content if the `direction` is `row`, or vertically aligns content if the `direction` is `column`.
 In the example below, `flex-start` will align the children content to the left.
 
 ##### wrap
 
-**Type**: `boolean`
+**Type**: `[boolean]`
 
 Determines if children should wrap.
