+++ packages/components/src/flex/flex/index.js
@@ -1,0 +1,2 @@
+export { default } from './component';
+export { useFlex } from './hook';
