+++ packages/components/src/flex/index.js
@@ -1,7 +1,3 @@
-export { default as Flex } from './flex';
-export { default as FlexItem } from './flex-item';
-export { default as FlexBlock } from './flex-block';
-
-export * from './use-flex';
-export * from './use-flex-item';
-export * from './use-flex-block';
+export { default as Flex, useFlex } from './flex';
+export { default as FlexItem, useFlexItem } from './flex-item';
+export { default as FlexBlock, useFlexBlock } from './flex-block';
